+++ example/AllMaterials/shaders/defines.glsl
@@ -1,0 +1,3 @@
+uniform float uSeed;
+uniform float uTime;
+uniform int uType;