+++ example/AllMaterials/shaders/frag/header.glsl
@@ -1,0 +1,2 @@
+
+varying vec3 vPosition;