+++ example/AllMaterials/shaders/frag/main.glsl
@@ -1,0 +1,1 @@
+vec4 newColor = vec4(sin(uTime), cos(uTime), sin(uTime * 0.5), 1.0);